migrate to monadic hooks, reorganize sequence components
