+++ modules/web/client/src/main/scala/observe/ui/components/sequence/steps/StepProgressCell.scala
@@ -41,7 +41,7 @@ case class StepProgressCell(
   progress:        Option[StepProgress],
   selectedStep:    Option[Step.Id],
   isPreview:       Boolean
-) extends ReactFnProps(StepProgressCell.component):
+) extends ReactFnProps(StepProgressCell):
   // def stepSelected(i: Step.Id): Boolean =
   //   selectedStep.exists(_ === i) && !isPreview &&
   //     // Will we always require logging in?
@@ -70,117 +70,115 @@ case class StepProgressCell(
   def isStopping: Boolean =
     requests.stop === OperationRequest.InFlight
 
-object StepProgressCell:
-  private type Props = StepProgressCell
+object StepProgressCell
+    extends ReactFnComponent[StepProgressCell](props =>
+      val stepControlButtons: TagMod =
+        StepControlButtons(
+          props.obsId,
+          props.instrument,
+          props.sequenceState,
+          props.stepId,
+          props.isPausedInStep,
+          props.progress.exists(_.stage === ObserveStage.ReadingOut),
+          false, // props.isNs,
+          props.requests
+        )        // .when(props.controlButtonsActive)
 
-  private def stepControlButtons(props: Props): TagMod =
-    StepControlButtons(
-      props.obsId,
-      props.instrument,
-      props.sequenceState,
-      props.stepId,
-      props.isPausedInStep,
-      props.progress.exists(_.stage === ObserveStage.ReadingOut),
-      false, // props.isNs,
-      props.requests
-    )        // .when(props.controlButtonsActive)
+        // def step.getSystemsStatus(step.get: Step): VdomElement =
+        //   <.div(
+        //     ObserveStyles.configuringRow,
+        //     <.div(
+        //       ObserveStyles.specialStateLabel,
+        //       "Configuring"
+        //     ),
+        //     <.div(
+        //       ObserveStyles.subsystems,
+        //       Step.configStatus
+        //         .getOption(step.get)
+        //         .orEmpty
+        //         .sortBy(_._1)
+        //         .map(Function.tupled(statusLabel))
+        //         .toTagMod
+        //     )
+        //   )
 
-  // private def step.getSystemsStatus(step.get: Step): VdomElement =
-  //   <.div(
-  //     ObserveStyles.configuringRow,
-  //     <.div(
-  //       ObserveStyles.specialStateLabel,
-  //       "Configuring"
-  //     ),
-  //     <.div(
-  //       ObserveStyles.subsystems,
-  //       Step.configStatus
-  //         .getOption(step.get)
-  //         .orEmpty
-  //         .sortBy(_._1)
-  //         .map(Function.tupled(statusLabel))
-  //         .toTagMod
-  //     )
-  //   )
+        // (props.sequenceState, props.step.get) match
+        //       case (f, s) if s.status === StepState.Running && s.fileId.isEmpty && f.userStopRequested =>
+        // Case pause at the sequence level
+        // step.getObservationPausing(props)
+        // EmptyVdom
+        //       case (_, s) if s.status === StepState.Running && s.fileId.isEmpty                        =>
+        // Case configuring, label and status icons
+        // step.getSystemsStatus(s)
+        // EmptyVdom
+        // case (_, s) if s.isObservePaused && s.fileId.isDefined                =>
+        // Case for exposure paused, label and control buttons
+        // stepObservationStatusAndFile(props, s.fileId.orEmpty, paused = true)
+        // case (_, s) if s.status === StepState.Running && s.fileId.isDefined   =>
+        // Case for a exposure onging, progress bar and control buttons
+        // stepObservationStatusAndFile(props, s.fileId.orEmpty, paused = false)
+        // case (_, s) if s.wasSkipped                                                              =>
+        //   <.p("Skipped")
+        // case (_, _) if props.step.get.skip                                                       =>
+        //   <.p("Skip")
+        // case (_, s) if s.status === StepState.Completed && s.fileId.isDefined =>
+        //   <.p(ObserveStyles.ComponentLabel, s.fileId.map(_.value).orEmpty)
+        // case (_, s) if props.stepSelected(s.id) && s.canConfigure             =>
+        //   // step.getSubsystemControl(props)
+        //   EmptyVdom
+        // case _                                                                =>
+        // <.p(ObserveStyles.ComponentLabel, props.step.get.shortName)
 
-  private val component = ScalaFnComponent[Props](props =>
-    // (props.sequenceState, props.step.get) match
-    //       case (f, s) if s.status === StepState.Running && s.fileId.isEmpty && f.userStopRequested =>
-    // Case pause at the sequence level
-    // step.getObservationPausing(props)
-    // EmptyVdom
-    //       case (_, s) if s.status === StepState.Running && s.fileId.isEmpty                        =>
-    // Case configuring, label and status icons
-    // step.getSystemsStatus(s)
-    // EmptyVdom
-    // case (_, s) if s.isObservePaused && s.fileId.isDefined                =>
-    // Case for exposure paused, label and control buttons
-    // stepObservationStatusAndFile(props, s.fileId.orEmpty, paused = true)
-    // case (_, s) if s.status === StepState.Running && s.fileId.isDefined   =>
-    // Case for a exposure onging, progress bar and control buttons
-    // stepObservationStatusAndFile(props, s.fileId.orEmpty, paused = false)
-    // case (_, s) if s.wasSkipped                                                              =>
-    //   <.p("Skipped")
-    // case (_, _) if props.step.get.skip                                                       =>
-    //   <.p("Skip")
-    // case (_, s) if s.status === StepState.Completed && s.fileId.isDefined =>
-    //   <.p(ObserveStyles.ComponentLabel, s.fileId.map(_.value).orEmpty)
-    // case (_, s) if props.stepSelected(s.id) && s.canConfigure             =>
-    //   // step.getSubsystemControl(props)
-    //   EmptyVdom
-    // case _                                                                =>
-    // <.p(ObserveStyles.ComponentLabel, props.step.get.shortName)
+        // if (props.isRunning)
+        //   stepObservationStatusAndFile(props, ImageFileId(""), paused = false)
+        // else
+        //   <.p(ObserveStyles.ComponentLabel, "Pending")
 
-    // if (props.isRunning)
-    //   stepObservationStatusAndFile(props, ImageFileId(""), paused = false)
-    // else
-    //   <.p(ObserveStyles.ComponentLabel, "Pending")
+      val fileId = ImageFileId("")
 
-    val fileId = ImageFileId("")
-
-    <.div(ObserveStyles.ConfiguringRow)(
-      stepControlButtons(props),
-      SubsystemControls(
-        props.obsId,
-        props.stepId,
-        props.subsystemStatus.map(_._1).toList,
-        props.subsystemStatus,
-        props.requests.subsystemRun.getOrElse(props.stepId, Map.empty),
-        props.sequenceState,
-        props.systemOverrides,
-        props.clientMode
-      ),
-      if (props.isBias)
-        // BiasStatus(
-        //   props.obsIdName,
-        //   props.step.get.id,
-        //   fileId,
-        //   stopping = !paused && props.isStopping,
-        //   paused
-        // )
-        EmptyVdom
-      else if (props.isRunning)
-        // props.nsStatus.fold[VdomNode] {
-        ObservationProgressBar(
+      <.div(ObserveStyles.ConfiguringRow)(
+        stepControlButtons,
+        SubsystemControls(
           props.obsId,
           props.stepId,
+          props.subsystemStatus.map(_._1).toList,
+          props.subsystemStatus,
+          props.requests.subsystemRun.getOrElse(props.stepId, Map.empty),
           props.sequenceState,
-          props.exposureTime,
-          props.progress,
-          fileId,
-          isStopping = !props.isPausedInStep && props.isStopping,
-          props.isPausedInStep
-        )
-      else EmptyVdom
-      // } { nsStatus =>
-      //   // NodAndShuffleProgressMessage(props.obsIdName,
-      //   //                              props.step.get.id,
-      //   //                              fileId,
-      //   //                              props.isStopping,
-      //   //                              paused,
-      //   //                              nsStatus
-      //   // )
-      //   EmptyVdom
-      // },
+          props.systemOverrides,
+          props.clientMode
+        ),
+        if (props.isBias)
+          // BiasStatus(
+          //   props.obsIdName,
+          //   props.step.get.id,
+          //   fileId,
+          //   stopping = !paused && props.isStopping,
+          //   paused
+          // )
+          EmptyVdom
+        else if (props.isRunning)
+          // props.nsStatus.fold[VdomNode] {
+          ObservationProgressBar(
+            props.obsId,
+            props.stepId,
+            props.sequenceState,
+            props.exposureTime,
+            props.progress,
+            fileId,
+            isStopping = !props.isPausedInStep && props.isStopping,
+            props.isPausedInStep
+          )
+        else EmptyVdom
+        // } { nsStatus =>
+        //   // NodAndShuffleProgressMessage(props.obsIdName,
+        //   //                              props.step.get.id,
+        //   //                              fileId,
+        //   //                              props.isStopping,
+        //   //                              paused,
+        //   //                              nsStatus
+        //   // )
+        //   EmptyVdom
+        // },
+      )
     )
-  )
